requirements.txt
psycopg2-binary
redis
Pillow